+++ aiobungie/api/rest.py
@@ -55,16 +55,14 @@ class RESTClient(traits.RESTful, abc.ABC):
 
     if typing.TYPE_CHECKING:
 
-        async def __aenter__(self) -> Self:
-            ...
+        async def __aenter__(self) -> Self: ...
 
         async def __aexit__(
             self,
             exception_type: type[BaseException] | None,
             exception: BaseException | None,
             exception_traceback: types.TracebackType | None,
-        ) -> None:
-            ...
+        ) -> None: ...
 
     @abc.abstractmethod
     async def read_manifest_bytes(self, language: str = "en", /) -> bytes:
@@ -2315,24 +2313,20 @@ class RESTClient(traits.RESTful, abc.ABC):
         """
 
     @abc.abstractmethod
-    async def fetch_historical_definition(self) -> typedefs.JSONObject:
-        ...
+    async def fetch_historical_definition(self) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
-    async def fetch_content_type(self, type: str, /) -> typedefs.JSONObject:
-        ...
+    async def fetch_content_type(self, type: str, /) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_content_by_id(
         self, id: int, locale: str, /, *, head: bool = False
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_content_by_tag_and_type(
         self, locale: str, tag: str, type: str, *, head: bool = False
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def search_content_with_text(
@@ -2345,8 +2339,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         *,
         page: int | None = None,
         source: str | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def search_content_by_tag_and_type(
@@ -2356,14 +2349,12 @@ class RESTClient(traits.RESTful, abc.ABC):
         type: str,
         *,
         page: int | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def search_help_articles(
         self, text: str, size: str, /
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_topics_page(
@@ -2376,8 +2367,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         page: int | None = None,
         locales: collections.Iterable[str] | None = None,
         tag_filter: str | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_core_topics_page(
@@ -2388,8 +2378,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         *,
         page: int | None = None,
         locales: collections.Iterable[str] | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_posts_threaded_page(
@@ -2402,8 +2391,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         root_thread_mode: bool,
         sort_mode: int,
         show_banned: str | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_posts_threaded_page_from_child(
@@ -2415,38 +2403,31 @@ class RESTClient(traits.RESTful, abc.ABC):
         root_thread_mode: bool,
         sort_mode: int,
         show_banned: str | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_post_and_parent(
         self, child_id: int, /, *, show_banned: str | None = None
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_posts_and_parent_awaiting(
         self, child_id: int, /, *, show_banned: str | None = None
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
-    async def fetch_topic_for_content(self, content_id: int, /) -> int:
-        ...
+    async def fetch_topic_for_content(self, content_id: int, /) -> int: ...
 
     @abc.abstractmethod
     async def fetch_forum_tag_suggestions(
         self, partial_tag: str, /
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
-    async def fetch_poll(self, topic_id: int, /) -> typedefs.JSONObject:
-        ...
+    async def fetch_poll(self, topic_id: int, /) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
-    async def fetch_recruitment_thread_summaries(self) -> typedefs.JSONArray:
-        ...
+    async def fetch_recruitment_thread_summaries(self) -> typedefs.JSONArray: ...
 
     @abc.abstractmethod
     async def fetch_recommended_groups(
@@ -2456,12 +2437,10 @@ class RESTClient(traits.RESTful, abc.ABC):
         *,
         date_range: int = 0,
         group_type: enums.GroupType | int = enums.GroupType.CLAN,
-    ) -> typedefs.JSONArray:
-        ...
+    ) -> typedefs.JSONArray: ...
 
     @abc.abstractmethod
-    async def fetch_available_avatars(self) -> collections.Mapping[str, int]:
-        ...
+    async def fetch_available_avatars(self) -> collections.Mapping[str, int]: ...
 
     @abc.abstractmethod
     async def fetch_user_clan_invite_setting(
@@ -2469,8 +2448,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         access_token: str,
         /,
         membership_type: enums.MembershipType | int,
-    ) -> bool:
-        ...
+    ) -> bool: ...
 
     @abc.abstractmethod
     async def fetch_banned_group_members(
@@ -2480,20 +2458,17 @@ class RESTClient(traits.RESTful, abc.ABC):
         /,
         *,
         page: int = 1,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_pending_group_memberships(
         self, access_token: str, group_id: int, /, *, current_page: int = 1
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def fetch_invited_group_memberships(
         self, access_token: str, group_id: int, /, *, current_page: int = 1
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def invite_member_to_group(
@@ -2505,8 +2480,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         membership_type: enums.MembershipType | int,
         *,
         message: str | None = None,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def cancel_group_member_invite(
@@ -2516,8 +2490,7 @@ class RESTClient(traits.RESTful, abc.ABC):
         group_id: int,
         membership_id: int,
         membership_type: enums.MembershipType | int,
-    ) -> typedefs.JSONObject:
-        ...
+    ) -> typedefs.JSONObject: ...
 
     @abc.abstractmethod
     async def equip_loadout(